requirements.txt
appdirs==1.4.4
asgiref==3.5.2
Babel==2.10.3
black==22.8.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.1
defusedxml==0.7.1
Django==3.2.7
django-autoslug==1.9.8
django-countries==7.2.1
django-environ==0.7.0
django-filter==21.1
django-phonenumber-field==2.2.0
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
flake8==3.9.2
idna==3.4
isort==5.9.3
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
mccabe==0.6.1
mypy-extensions==0.4.3
oauthlib==3.2.1
pathspec==0.10.1
phonenumbers==8.12.33
Pillow==8.3.2
platformdirs==2.5.2
psycopg2-binary==2.9.1
pycodestyle==2.7.0
pycparser==2.21
pyflakes==2.3.1
PyJWT==2.5.0
python3-openid==3.2.0
pytz==2022.2.1
regex==2022.9.13
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
toml==0.10.2
tomli==2.0.1
uritemplate==4.1.1
urllib3==1.26.12
